% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_data}
\alias{env_data}
\title{Environmental data by year, month, and location}
\format{A data frame with 24288 rows and 4 variables:
\describe{
  \item{month}{as integer}
  \item{year}{as integer}
  \item{envVar}{An environmental variable, standardized by month}
  \item{Location}{character, label for location}
}}
\source{
Created by the package author.
}
\usage{
env_data
}
\description{
A anonymized dataset containing an environmental covariate measured by month at many
locations for several years
}
\keyword{datasets}
