% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flmtools-package.R
\docType{package}
\name{flmtools-package}
\alias{flmtools}
\alias{flmtools-package}
\title{flmtools: Utilities for Working with Functional Linear Models}
\description{
Functional Linear Models allow matrix covariates, but arranging the data and working with the outputs is difficult. These functions aim to make life easier, mostly for the author.
}
\author{
\strong{Maintainer}: Andrew Tyre \email{atyre2@unl.edu}

}
\keyword{internal}
